User prompts to help the Mark 1 get connected to the internet didn't always occur
On the Mark 1 device, there is a sequence of events that produce a guide to the user for getting their device setup and connected to the internet.  This process is basically:
* Enclosure client starts up, sending a message to the Arduino
* When/if the Arduino responds, it checks for a live internet connection
* If no connection is found, the user is prompted
However the speech client was sometimes not fully up, so the request to speak the prompt was sent before it was listening.

Additionally, sometimes events would occur that caused the prompt to connect the unit multiple time in a row.  This commit limits the prompts to a max of once every 30 seconds.

Finally, I commented the code and renamed some of the methods to clarify this sequence of events.  The code:
   Timer(5, self.stop).start()
and:
   def stop(self):
        if not self.started:
            self.writer.stop()
Are some of the most confusing lines of code I've ever seen.  :)

Major changes:
* The messagebus message "enclosure.start" is now "enclosure.started"
* Enclosure.start() is now Enclosure.on_arduino_responded()
* Enclosure.started is now Enclosure.arduino_responded
* Added Enclosure.last_internet_notification to track notification times
* The notification call to the speech client is now delayed by 5 seconds, providing time for the speech client to come up.  (And an additional check is run to verify the connection is still down immediately before the verbal prompt is spoken)