Commit headline -- not longer than this line (50)
==== Fixed Issues ====
984 - install_mimic.sh -- Mimic shouldn't have to clone all branches

====  Tech Notes ====
I made a small edit to the install-mimic.sh script to clone the latest
version branch with the --depth=1 option instead of a full clone, this
saves a bit of download time and space

==== Environment Notes ====
This may cause issues with people if they want to change to a different
mimic branch. This is becasue only a single branch will be fetched
locally.