Fix MSM permissions installing on Mark 1/Picroft
On a Raspbian platform, MSM was attempting to perform a 'sudo chown'
on the skill folder after the install using a sudo command.  When
running inside mycroft-core as the 'mycroft' user, there is no
'sudo' permission.  However, this step isn't even needed since the
user installed as mycroft and the folder has the correct permissions.
Now it checks the user and group before attempting to perform the
action.