Disconnect test backend from unstable in mycroft-use
The mycroft-use utility had a direct connection between the unstable build and
the test backend; and the stable build and the production backend.  This made
some kinds of testing tricky.

Now you run ```mycroft-use home``` and ```mycroft-use home-test``` to make
the backend association explicitly.

Also cleaned up the whitespace, which was a mixture of spaces and tabs.