Add function to get configuration value and use it
The get_config_value function takes two arguments "key" and "default
value" and looks for the key to be defined in ~/.mycroft/mycroft.conf
and /etc/mycroft/mycroft.conf. First looking into the user defined
file and falling back to the system-wide configuration file.

If the key can't be found in any of both places, it returns the default
value given as argument

Use the new function to get the values for mycroft_platform and
mycroft_skill_folder