Move the /opt/mycroft/web_config_cache.json file to ~/.mycroft/cache/
The /opt directory is a system folder that in some systems is not
writable. It's better to not assume it is (or force it being writable)
but use a directory that we know it's assured to be writable, like
~/.mycroft .

This helps mycroft being able to be used by regular desktop users which
this way would each use their own cache files.