Added a new feature to... nevermind
The utterance processing now looks at the tail end of the transcribed phrase, ignoring the utterance if it ends with "nevermind" or "cancel" or "ignore that".  This is for use case like:
   Hey Mycroft, can you tell me the...ummm...oh, nevermind

I have some concerns about a few corner cases that a skill might want to use.
For example:
 * Hey Mycroft, play Nirvana's Nevermind
 * Hey Mycroft, I want that last request to cancel
But I think this is something that can normally be designed around easily.

====  Tech Notes ====
Added mycroft.dialog.get_all(phrase, lang, context).  This lets you load all of the rendered versions, not just a single randomly selected one.

This class has the experiment of returning a value when a matching .dialog file isn't found.  Instead it returns ['phrase'], replacing the . with spaces.  So if I add:
    mycroft.dialog.get_all("thank.you") and there is no resource file found named 'thank.you.dialog', it will still return successfully:  ['thank you']

==== Localization Notes ====
This adds the new "nevermind" core resource string.