Add support for 'destination' in message context
The destination adds one or multiple targets for an interaction. An output such as the audio subsystem will only output the text if the provided destination matches the subsytem. The idea is that a new client such as a web browser will set the destination context making sure the response isn't spoken and gives the ability to handle multple parallell sessions where the query of one user only results in an anwer to that user.

The 'destination' field should be a list but a single string will also work. if the field is missing or set to None the message is considered a multicast and should be spoken by all clients.

The Audio will disregard any speak messages with a 'destination' other than 'audio or None.