Run pip after any update to skills managed by msm.
Run pip after any updates to skills, to ensure that libraries are
updated per the lateset version of the skill. This may cause pip to run
twice (if mycroft has re-PIPed), however, given that the cost of this
is very low (basically a no-op on the second run), and the re-PIP logic
is flagged for removal, it seems safest to not touch that logic for
now.

==== Fixed Issues ====