Fix install script for renamed module
This fixes a problem where if "mycroft-core" was already in the
virtualenv path file, but the actual name of the module wasn't,
then the dev_setup.sh would succeed but start-mycroft.sh would
fail with "No module named mycroft.version". Now dev_setup.sh
actually checks for the $TOP variable in the virtualenv path,
instead of hardcoded "mycroft-core".

==== Fixed Issues ====
Might fix issue #1419 for some rare cases.

====  Documentation Notes ====
Slightly improves the durability of the install script.