Add 'venv-activate.sh' script (#1682)
Add 'venv-activate.sh' script

This script in the root makes it easier for devs to discover the
virtual environment which Mycroft should run under.  Virtual
environments are new to many, so the commands aren't obvious even
if they are only one liners.

====  Documentation Notes ====
Should probably be mentioned in the Getting Started docs for
Github users.  This mirrors the 'mycroft-venv-activate' alias found
under Mark 1/Picroft.